requirements.txt
uwsgi
Django==2.0.7
elasticsearch-dsl
django-elasticsearch-dsl
djangorestframework==3.8.2
django-elasticsearch-dsl-drf==0.11
ipython
